Remove class on click to prevent lingering hover effect
The CSS :hover selector requires a mouseleave/mouseout to remove its
effect, and because of the sliding top bar the mouseleave doesn't
happen on touch devices.

By adding and removing a "hover" class name, the JS can remove the
"hover" on click as well to prevent having to wait for a mouse out.

Fixes #7.